Looting and lots of coments.
Added a bunch of comments and added looting of ships! Player can not
disable ships and loot them for cargo and money! Added new type of ship,
"wreck" that inherets much from its origional but does not move or fire
and can be looted. Added looting UI. Curently UI code has lots of copy
pastes, is this a prob, maybe...