Massive (way to big) changes
This update was way more than i could chew in one update, thats why
took 2 months. Major changes to how Items and Ships work.
Ships now have 2 "components" the ship itself with physical things, and
the ai which determines behavior.
items can be active or passive or not equipment at all
Guns and trade goods now create genirators which then create actual
pieces of equipment,
UI for slots has been started but not yet finished
THIS UPDATE IS UNSTABLE AND BAD, but it is headed the right direction.
Gona be bugs galore. Take this as a lesson that fixing things piece by
piece is easyer than all at once. GAAAAAA